requirements.txt
Flask
pandas
requests
Flask-Session
pywebview>=5.0
